feat:Customer-> Mark as rented
