+++ app/controllers/API/Orders.php
@@ -95,20 +95,6 @@ class Orders {
 
     // Customer
 
-    public function markAsRentedByCustomer(string $a = '', string $b = '', string $c = ''):void {
-
-        $response = new JSONResponse();
-
-        $data = [
-            'customer_req' => 'rented',
-        ];
-
-        $rentReq = new RentRequestModel;
-        $rentReq->updateRequest($a, $data);
-
-        $response->statusCode(200)->data(['order_id' => $a])->send();
-
-    }
 
     public function markAsReturnedByCustomer(string $a = '', string $b = '', string $c = ''):void {
